Remove the "c" UOM of perfdata output. Closes #30.
The issue has been introduced in 3e27344 by issue #18.

The "c" thing means it's a continous counter (a number increasing over
time). However the output of check_iftraffic_nrpe.py is not a counter
but a value of bytes per second (by default).

The Nagios plugins guideline does not specify a suitable unit but "no
unit specified".